feat: bottomNavigationBar that replaces app drawer
initially, each link in the app drawer returns a different Scaffold, using the MaterialApp’s Navigator to push to different routes.

however, this is not good practice as we were nesting a Scaffold within a Scaffold. additionally, the state of the bottomNavigationBar was lost every time a link was clicked.

`NavigatorMain` is a Scaffold that has a `Navigator` in its `body` property. a key is passed to both the `NavBar` and `Navigator`.

problems:
- “Null check operator used on a null value” error upon removing `home` property from `MaterialApp`, despite removing all bang operators that could have caused the error.
- registering signs a user in automatically
- map screen doesn’t work, as we need a floating button that is provided by Scaffold, but we cannot `setState` of `NavigatorMain`’s floating button in `MapWidget` after it has been built.

todo:
- settings page (currently, logs user out on click)
- add email verification after registration (which would also prevent user from being signed in automatically after registration)

other learning points:
- routes between different navigators are not additive (i.e. if you have a navigator with route ‘/‘ and a nested one with route ‘add’, the route now is NOT ‘/add’)
- an initialRoute prefixed by ‘/‘ will load the ‘/‘ route first; thus an error will occur if the ‘/‘ route is not defined