feat: add widget with existing stalls and locations
workflow:
1. user selects from existing stalls in the firestore db, or creates a new stall
2. user types in a new location for the stall, or selects from the existing locations
3. user gives the rating and submits

- algolia is used to provide search and filtering capabilities to get the list of existing stall
names and the list of existing locations for a selected stall
- `showSearch` function is used to display the search results for the existing stall names

there were two main problems that required a redesign of the db structure each time.
1. inability to edit an item in an array in firestore.:%s/.\{80}/&\r/g
2. inability to use a single index for the subcollections of all documents in a collection (e.g. if
stalls is a collection, the path name would be stalls/*/*)

originally, the plan was to have a collection of stalls, with each document containing this
structure:
```
Name:
Branches: [{
	Location:
	Reviews: [{
		Rating:
		User:
	}]
}]
```
however, because firestore only allows us to add or delete items in an array, it is not possible to
edit the ‘reviews’ array without first deleting the whole map for a branch and then adding it back
with the new review, which would not be efficient if there are many reviews.

the next solution was to use a subcollection for `Branches` instead of an array. the new structure
for each stall document:
```
Name:
Coordinates:
(Collection) branches:

// document in branches
Location:
Reviews: [{
	Rating:
	User:
}]

```
this would mean that every location for a stall would be represented by a document in the `Branches`
 subcollection of the stall document. however, we would need to add a new index in algolia for every
 of such document, and also install a new instance of `Search with Algolia` extension in Firebase to
 update this index.

this seems inefficient and time-consuming, especially since every new instance of `Search with
Algolia` takes around 3-5 minutes to create and must be done so manually.

the final solution uses a separate ‘reviews’ collection. each document in this collection has the
document id of ‘{stall_name}; {location_of_stall}’. the new structures:

```
// document in stalls
Names:
Locations: []

// document in reviews
Reviews: [{
	Rating:
	UserPath:
	Date Added:
}]
```
we maintain the ability to obtain a list of stall names, and the ability to obtain a list of
locations for an existing stall from the stalls collection.

we also avoid having to add new indexes as the number of stalls and locations increases (only 2
indexes, for the stalls and reviews collections respectively).